Bulk Save Documentation
The session::bulk_save_objects method emphasizes performance, but
generates many insert statements when a variety of object types are
passed in an unordered list. This was a signification gotcha for me,
and seems contrary to the spirit of the method, so short of changing
the method to automatically sort the list, it should at least be
documented.