Clarify mappers.configure_mappers error message
Sometimes the message generated by configure_mappers lacks a reference
to the failing mapper, and only the mapper._configure_failed is shown. This
makes debugging problems extremely difficult. Example:

    sqlalchemy.exc.InvalidRequestError: One or more mappers failed to
    initialize - can't proceed with initialization of other mappers.  Original
    exception was: Class 'neutron.objects.router.Router' is not mapped

In the above failure, the actual object having a problem is FloatingIP, which
has a knock-on effect on the Router object when it fails to map. A more helpful
error message might look like this example:

    sqlalchemy.exc.InvalidRequestError: One (Mapper|FloatingIp|floatingip) or
    more mappers failed to initialize - can't proceed with initialization of
    other mappers.  Original exception was: Class
    'neutron.objects.router.Router' is not mapped

This patch adds the failing mapper name to the beginning of the message after
'One', as shown in the second example.

Change-Id: I9f23bfa90b26dde9229ab7ec812eec9ceae48153