Fix postgres 9.5 operator precedence change
Postgres 9.5 subtly changed the operator precedence, which breaks some
existing code. In this patch I change the operator precedence to be
inline with postgres's new operator precedence, which results in more
operators sharing a precedence. This will mean that more brackets are
generated than before, which should be benign on other dbs.
Also a unit test for the case which broke for me. And updated a couple
of other unit tests to pass with extra brackets.