Fix lazy loading of relations from identity map for sharded setups
After shipping 1.2 to production, we noticed a significant increase in
query throughput. After investigation, I think the loading strategy
needs to supply identity_token to identity_key_from_primary_key in order
for there to be a successful lookup against the identity map.

Is it safe to pass the state's identity token?