I adjusted the copy-and-paste warning wording.
I found the use of the word "left" in this sentence to be slightly
confusing.

  Ignoring declarative-like tuple value of attribute %s: possibly
  a copy-and-paste error with a comma left at the end of the line?

The sentence is entirely correct, using the sense of left as
"remaining", but I initially read it as the comma being at the
left side of the line, that is, the beginning rather than the end.

I suggest changing the word "left" to anything else, or removing it.
The present commit changes it to "accidentally". Other options include
"remaining", "still", or no word.